@@ -21,16 +21,16 @@
   },
   "license": "MIT",
   "engines": {
-    "node": ">=6.0.0"
+    "node": ">=14.0.0"
   },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com/"
   },
   "scripts": {
-    "lint": "tslint --project tsconfig.json -t codeFrame 'lib/src/**/*.ts' 'test/**/*.ts'",
-    "format": "prettier --write \"lib/src/**/*.ts\"",
+    "lint": "eslint . --ext .ts",
+    "format": "prettier --write .",
     "prebuild": "rimraf dist",
-    "build": "tsc --module commonjs && rollup -c rollup.config.ts && typedoc --out docs --target es6 --theme minimal --mode file lib",
+    "build": "tsc --module commonjs && rollup -c rollup.config.ts && typedoc lib/bloc.ts --out docs --theme default",
     "start": "rollup -c rollup.config.ts -w",
     "test": "jest --coverage",
     "test:watch": "jest --coverage --watch",
@@ -38,6 +38,7 @@
   },
   "lint-staged": {
     "{src,test}/**/*.ts": [
+      "eslint --fix",
       "prettier --write",
       "git add"
     ]
@@ -76,13 +77,17 @@
   "devDependencies": {
     "@types/jest": "^23.3.2",
     "@types/node": "^10.14.15",
+    "@typescript-eslint/eslint-plugin": "^5.51.0",
+    "@typescript-eslint/parser": "^5.51.0",
     "colors": "^1.3.2",
     "cross-env": "^5.2.0",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
     "jest": "^26.4.2",
     "jest-config": "^26.4.2",
     "lint-staged": "^13.1.0",
     "lodash.camelcase": "^4.3.0",
-    "prettier": "^1.14.3",
+    "prettier": "^2.8.4",
     "replace-in-file": "^3.4.2",
     "rimraf": "^2.6.2",
     "rollup": "^0.67.0",
@@ -94,11 +99,8 @@
     "rxjs": "^6.5.3",
     "ts-jest": "^26.3.0",
     "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.15.0",
-    "tslint-config-standard": "^8.0.1",
-    "typedoc": "^0.15.0",
-    "typescript": "^3.0.3"
+    "typedoc": "^0.23.24",
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
     "rxjs": "6.x.x"
