@@ -1,225 +1,167 @@
-import { Observable, Subject, EMPTY, Subscription } from 'rxjs'
-import { catchError, concatMap, map } from 'rxjs/operators'
-import { BlocObserver, EventStreamClosedError, Transition } from '../bloc'
+import { EMPTY, from, Observable, Subject, Subscription } from 'rxjs'
 
-export type NextFunction<Event, State> = (value: Event) => Observable<Transition<Event, State>>
+import { catchError, filter, mergeMap } from 'rxjs/operators'
+import { BlocBase } from './base'
+import { BlocObserver } from './bloc-observer'
+import { Emitter, _Emitter } from './emitter'
+import { StateError } from './errors'
+import { BlocEvent } from './event'
+import { Transition } from './transition'
+import { EventHandler, ClassType, EventTransformer } from './types'
 
-/**
- *
- * Takes a stream of `Events` as input and transforms them into a Stream of `States` as output.
- *
- * @export
- * @abstract
- * @class Bloc
- * @extends {Observable<State>}
- * @template Event
- * @template State
- */
-export abstract class Bloc<Event, State> extends Observable<State> {
-  constructor(private _state: State) {
-    super()
-    this.stateSubject = new Subject()
-    this.bindStateSubject()
+export abstract class Bloc<Event extends BlocEvent, State> extends BlocBase<State> {
+  constructor(state: State) {
+    super(state)
+    this.on = this.on.bind(this)
+    this.add = this.add.bind(this)
+    this.emit = this.emit.bind(this)
   }
 
-  private emitted: boolean = false
-  private eventSubject = new Subject<Event>()
-  private stateSubject: Subject<State>
-  private transitionSubscription: Subscription = Subscription.EMPTY
+  private readonly _eventSubject$ = new Subject<Event>()
+  private readonly _eventMap = new WeakMap<ClassType<Event>, 1>()
+  private readonly _subscriptions = new Set<Subscription>()
+  private readonly _emitters = new Set<Emitter<State>>()
+  readonly isBlocInstance = true
 
-  /**
-   * The current `BlocObserver`.
-   *
-   * @static
-   * @type {BlocObserver}
-   * @memberof Bloc
-   */
-  static observer: BlocObserver = new BlocObserver()
-
-  /**
-   * Returns the current state of the bloc.
-   *
-   * @readonly
-   * @type {State}
-   * @memberof Bloc
-   */
-  get state(): State {
-    return this._state
+  static transformer<T>(): EventTransformer<T> {
+    return (events$, mapper) => events$.pipe(mergeMap(mapper))
   }
 
-  /**
-   * Adds a Subscription to the bloc's state stream.
-   *
-   * @param {(value: State) => void} onData
-   * @param {(((onError: any) => any) | undefined)} [onError]
-   * @param {((() => any) | undefined)} [onDone]
-   * @return {*}  {Subscription}
-   * @memberof Bloc
-   */
-  listen(
-    onData: (value: State) => void,
-    onError?: ((onError: any) => any) | undefined,
-    onDone?: (() => any) | undefined
-  ): Subscription {
-    return this.stateSubject.subscribe(onData, onError, onDone)
+  private _mapEventToStateError(error: Error): Observable<never> {
+    this.onError(error)
+    return EMPTY
   }
 
-  /**
-   * Notifies the bloc of a new event which triggers `mapEventToState`.
-   *
-   * @param {Event} event
-   * @memberof Bloc
-   */
-  add(event: Event): void {
-    try {
-      if (this.eventSubject.isStopped) {
-        throw new EventStreamClosedError()
-      }
-      this.onEvent(event)
-      this.eventSubject.next(event)
-    } catch (error) {
-      this.onError(error)
-    }
-  }
-
-  /**
-   * Called whenever an event is added to the bloc.
-   *
-   * @param {Event} event
-   * @memberof Bloc
-   */
-  onEvent(event: Event): void {
-    Bloc.observer.onEvent(this, event)
-  }
-
-  /**
-   * Transforms the events along with a `NextFunction` into
-   * an `Observable<Transition>`.
-   * Events that should be processed by `mapEventToState` need to be passed to
-   * the `next`.
-   * By default `concatMap` is used to ensure all events are processed in
-   * the order in which they are received.
-   * You can override `transformEvents` for advanced usage in order to
-   * manipulate the frequency and specificity with which `mapEventToState` is
-   * called as well as which `events` are processed.
-   *
-   * @param {Observable<Event>} events
-   * @param {NextFunction<Event, State>} next
-   * @return {*}  {Observable<Transition<Event, State>>}
-   * @memberof Bloc
-   */
-  transformEvents(
-    events: Observable<Event>,
-    next: NextFunction<Event, State>
-  ): Observable<Transition<Event, State>> {
-    return events.pipe(concatMap(next))
-  }
-
-  /**
-   * Must be implemented when a class extends `Bloc`.
-   * Called whenever an event is added to the bloc and
-   * is responsible for converting incoming events into outgoing states.
-   *
-   * @abstract
-   * @param {Event} event
-   * @return {*}  {AsyncIterableIterator<State>}
-   * @memberof Bloc
-   */
-  abstract mapEventToState(event: Event): AsyncIterableIterator<State>
-
-  /**
-   * Transforms the `Observable<Transition>` into a new `Observable<Transition>`.
-   * By default `transformTransitions` returns the incoming `Observable<Transition>`.
-   * You can override `transformTransitions` for advanced usage in order to
-   * manipulate the frequency and specificity at which `transitions`
-   * (state changes) occur.
-   *
-   * @param {Observable<Transition<Event, State>>} transitions
-   * @return {*}  {Observable<Transition<Event, State>>}
-   * @memberof Bloc
-   */
-  transformTransitions(
-    transitions: Observable<Transition<Event, State>>
-  ): Observable<Transition<Event, State>> {
-    return transitions
-  }
-
-  /**
-   * Called whenever a `transition` occurs with the given `transition`.
-   * A `transition` occurs when a new `event` is added and `mapEventToState` executed.
-   * `onTransition` is called before a bloc's state has been updated.
-   *
-   * @param {Transition<Event, State>} transition
-   * @memberof Bloc
-   */
-  onTransition(transition: Transition<Event, State>): void {
-    Bloc.observer.onTransition(this, transition)
-  }
-
-  /**
-   * Called whenever an `error` is thrown within `mapEventToState`.
-   * By default all errors will be ignored and bloc functionality will be unaffected.
-   *
-   * @param {*} error
-   * @memberof Bloc
-   */
-  onError(error: any): void {
+  protected override onError(error: Error): void {
     Bloc.observer.onError(this, error)
   }
 
-  /**
-   * This method should be called when a `Bloc` is no longer needed.
-   * Disposes the resources held by the bloc which means the `Bloc` will
-   * no longer process new events after `close` has been called.
-   *
-   * @memberof Bloc
-   */
-  close(): void {
-    this.stateSubject.complete()
-    this.eventSubject.complete()
-    this.transitionSubscription.unsubscribe()
+  protected onTransition(transition: Transition<Event, State>): void {
+    Bloc.observer.onTransition(this, transition)
   }
 
-  private bindStateSubject(): void {
-    this.transitionSubscription = this.transformTransitions(
-      this.transformEvents(this.eventSubject, (event: Event) => {
-        return asyncToObservable(this.mapEventToState(event)).pipe(
-          map((nextState: State, _: number) => {
-            return new Transition(this.state, event, nextState)
-          }),
-          catchError(error => {
-            this.onError(error)
-            return EMPTY
-          })
-        )
-      })
-    ).subscribe((transition: Transition<Event, State>) => {
-      if (transition.nextState === this.state && this.emitted) return
-      try {
-        this.onTransition(transition)
-        this._state = transition.nextState
-        this.stateSubject.next(transition.nextState)
-      } catch (error) {
-        this.onError(error)
+  protected onEvent(event: Event): void {
+    Bloc.observer.onEvent(this, event)
+  }
+
+  protected on<T extends Event>(
+    event: ClassType<T>,
+    eventHandler: EventHandler<T, State>,
+    transformer?: EventTransformer<T>
+  ): void {
+    if (this._eventMap.has(event)) {
+      throw new Error(`${event.name} can only have one EventHandler`)
+    }
+
+    this._eventMap.set(event, 1)
+
+    const mapEventToState = (event: T): Observable<void> => {
+      const stateToBeEmittedStream$ = new Subject<State>()
+      let isClosed = false
+
+      const onEmit = (nextState: State): void => {
+        if (isClosed) {
+          return
+        }
+
+        if (this.state === nextState && this._emitted) return
+
+        try {
+          this.onTransition(new Transition(this.state, event, nextState))
+          stateToBeEmittedStream$.next(nextState)
+        } catch (error) {
+          this.onError(error as Error)
+        }
       }
-      this.emitted = true
-    })
+
+      const _emitter = new _Emitter(onEmit.bind(this))
+
+      const _callableEmitter = (state: State) => _emitter.call(state)
+
+      _callableEmitter.close = () => {
+        isClosed = true
+        _emitter.close()
+      }
+
+      _callableEmitter.onEach = <T>(
+        stream$: Observable<T>,
+        onData: (data: T) => void,
+        onError?: (error: Error) => void
+      ) => _emitter.onEach(stream$, onData, onError)
+
+      _callableEmitter.forEach = <T>(
+        stream$: Observable<T>,
+        onData: (data: T) => State,
+        onError?: (error: Error) => State
+      ) => _emitter.forEach(stream$, onData, onError)
+
+      this._emitters.add(_callableEmitter)
+
+      return new Observable((subscriber) => {
+        stateToBeEmittedStream$.subscribe(this.emit)
+
+        const result = eventHandler.call(this, event, _callableEmitter)
+
+        if (result instanceof Promise) {
+          from(result).subscribe({
+            complete: () => subscriber.complete(),
+          })
+        } else {
+          subscriber.complete()
+        }
+
+        return () => {
+          _callableEmitter.close()
+          this._emitters.delete(_callableEmitter)
+          stateToBeEmittedStream$.complete()
+        }
+      })
+    }
+
+    const _transformer = transformer ?? Bloc.transformer()
+
+    const transformStream$ = _transformer(
+      this._eventSubject$.pipe(filter((newEvent): newEvent is T => newEvent instanceof event)),
+      mapEventToState
+    )
+
+    const subscription = transformStream$
+      .pipe(catchError((error: Error) => this._mapEventToStateError(error)))
+      .subscribe()
+
+    this._subscriptions.add(subscription)
+  }
+
+  static observer: BlocObserver = new BlocObserver()
+
+  add(event: Event) {
+    if (!this._eventMap.has(event.constructor as ClassType<Event>)) {
+      throw new StateError(`
+        add(${event.blockEventName}) was called without a registered event handler.
+        Make sure to register a handler via on(${event.blockEventName}, (event, emit) {...})
+      `)
+    }
+
+    try {
+      this.onEvent(event)
+      this._eventSubject$.next(event)
+    } catch (error) {
+      this.onError(error as Error)
+    }
+
+    return this
+  }
+
+  override close(): void {
+    this._emitters.forEach((emitter) => emitter.close())
+    this._subscriptions.forEach((subscription) => subscription.unsubscribe())
+    this._emitters.clear()
+    this._subscriptions.clear()
+    super.close()
   }
 }
 
-function asyncToObservable<T>(iterable: AsyncIterableIterator<T>): Observable<T> {
-  return new Observable<T>(
-    observer =>
-      void (async () => {
-        try {
-          for await (const item of iterable) {
-            if (observer.closed) return
-            observer.next(item)
-          }
-          observer.complete()
-        } catch (e) {
-          observer.error(e)
-        }
-      })()
-  )
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const isBlocInstance = (bloc: any): bloc is Bloc<any, any> => {
+  return bloc instanceof Bloc || Boolean(bloc.isBlocInstance)
 }
